+++ CHANGELOG.md
@@ -10,7 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - **GitHub** - Workflow to run unittests on every PR and push to main.
 
 ### Changed
+- **Common** - Move EC2 pricing calls to single function in `common.py`.
 - **Readme** - Added new badges to the Readme
+
 
 ## [1.0.1] - 2022-09-28